+++ front/src/components/Select/Select.js
@@ -1,7 +1,7 @@
 import React, { useState, useEffect } from "react"
 import axios from "axios"
 
-import Table from "./Table/Table.js"
+import TableData from "./Table/TableData.js"
 
 function Select() {
   const [userId, setUserId] = useState("")
@@ -10,20 +10,12 @@ function Select() {
   const [grouplist, setGrouplist] = useState([])
   const [ingrelist, setIngrelist] = useState([])
   const [selected_group_value, setSelectedGroupValue] = useState("")
-  const [selected_ingre, setSelectedIngre] = useState([])
   let get_group_data = []
   let group_data_list = []
   let get_ingre_data = []
   let ingre_data_list = []
 
-  const getSelectedGroupValue = (selected_group_value) => {
-    setSelectedGroupValue(parseInt(selected_group_value))
-  }
-
-  const getSelectedIngre = (selected_ingre) => {
-    setSelectedIngre(selected_ingre)
-  }
-
+  // 유저의 못먹는 재료 리스트 불러오기
   useEffect(() => {
     setUserId(window.localStorage.getItem("userId"))
     setUserIdx(window.localStorage.getItem("idx"))
@@ -42,6 +34,7 @@ function Select() {
     userIdx !== undefined ? getInedible() : console.log("비로그인 상태")
   }, [userIdx])
 
+  // 재료군 불러오기
   useEffect(() => {
     function getGroup() {
       axios
@@ -60,6 +53,12 @@ function Select() {
     getGroup()
   }, [])
 
+  // 선택된 재료군 idx 불러오기
+  const getSelectedGroupValue = (selected_group_value) => {
+    setSelectedGroupValue(parseInt(selected_group_value))
+  }
+
+  // 선택된 재료군에 속한 재료들 불러오기
   useEffect(() => {
     function getSub() {
       axios
@@ -84,14 +83,13 @@ function Select() {
 
   return (
     <div>
-      <Table
+      <TableData
         userId={userId}
         userIdx={userIdx}
         inedible={inedible}
         grouplist={grouplist}
         getSelectedGroupValue={getSelectedGroupValue}
         ingrelist={ingrelist}
-        getSelectedIngre={getSelectedIngre}
       />
     </div>
   )
